Up to snuff. Prepare the hounds.
Respond to hound and drop in coverage

More responses to hound

Hound is hitting the github API too much, so have to fix continuously

more responses to hound

More responses to hound

Fixed typo

responses to hound

More responses to hound

+++ app/mutations/gardens/update_garden.rb
@@ -12,7 +12,7 @@ module Gardens
           string :type
           string :average_sun
           string :soil_type
-          integer :ph
+          float :ph
           array :growing_practices
         end
       end
